[poc/LEMS-3271-update-perseus-storybook] moved existing stories to be part of new structure
