+++ __docs__/perseus-components/components-overview.mdx
@@ -19,5 +19,3 @@ Perseus includes various reusable components such as:
 - **Tex** - LaTeX math rendering
 - **Zoomable** - Components with zoom functionality
 - **And many more...**
-
-These components are designed to be reusable, accessible, and consistent with Khan Academy's design principles.
