+++ __docs__/perseus-components/info-tip.stories.tsx
@@ -1,6 +1,6 @@
 import * as React from "react";
 
-import InfoTip from "../info-tip";
+import InfoTip from "../../packages/perseus/src/components/info-tip";
 
 import type {Meta, StoryObj} from "@storybook/react-vite";
