+++ __docs__/perseus-components/math-input.stories.tsx
@@ -1,6 +1,6 @@
 import {action} from "storybook/actions";
 
-import MathInput from "../math-input";
+import MathInput from "../../packages/perseus/src/components/math-input";
 
 import type {Meta, StoryObj} from "@storybook/react-vite";
