+++ __docs__/perseus-components/zoomable-tex.stories.tsx
@@ -1,6 +1,6 @@
 import * as React from "react";
 
-import ZoomableTex from "../zoomable-tex";
+import ZoomableTex from "../../packages/perseus/src/components/zoomable-tex";
 
 import type {Meta, StoryObj} from "@storybook/react-vite";
