+++ __docs__/perseus-widgets/perseus-radio-new/radio-editor.stories.tsx
@@ -2,7 +2,7 @@
 import {StyleSheet, css} from "aphrodite";
 import * as React from "react";
 
-import RadioEditor from "../../packages/perseus-editor/src/widgets/radio/editor";
+import RadioEditor from "../../../packages/perseus-editor/src/widgets/radio/editor";
 
 import type {APIOptions} from "@khanacademy/perseus";
 import type {
@@ -17,6 +17,7 @@ type Story = Meta<RadioEditor>;
 
 export default {
     title: "Perseus/Widgets/RadioNew/Editor Demo",
+    tags: ["!autodocs"],
 } as Story;
 
 const styles = StyleSheet.create({
