+++ __docs__/perseus-widgets/perseus-radio-new/radio-regression.stories.tsx
@@ -1,14 +1,14 @@
 import {generateTestPerseusItem} from "@khanacademy/perseus-core";
 import * as React from "react";
 
-import {ServerItemRendererWithDebugUI} from "../../testing/server-item-renderer-with-debug-ui";
+import {ServerItemRendererWithDebugUI} from "../../../testing/server-item-renderer-with-debug-ui";
 import {
     questionWithPassage,
     multiChoiceQuestion,
     choicesWithImages,
-} from "../../packages/perseus/src/widgets/radio/__tests__/radio.testdata";
+} from "../../../packages/perseus/src/widgets/radio/__tests__/radio.testdata";
 
-import type {APIOptions} from "../../packages/perseus/src/types";
+import type {APIOptions} from "../../../packages/perseus/src/types";
 import type {PerseusItem} from "@khanacademy/perseus-core";
 import type {Meta} from "@storybook/react-vite";
