+++ __docs__/perseus-widgets/perseus-radio-new/radio-states-gallery.stories.tsx
@@ -8,6 +8,7 @@ export default {
     parameters: {
         viewMode: "docs",
     },
+    tags: ["!autodocs"],
 } as Meta;
 
 const styles = StyleSheet.create({
